Added logic for swapping courses
