+++ app/src/main/java/com/example/peterchu/watplanner/home/HomePresenter.java
@@ -24,6 +24,7 @@ import com.example.peterchu.watplanner.data.DataRepository;
 import com.example.peterchu.watplanner.data.IDataRepository;
 import com.example.peterchu.watplanner.scheduler.CourseScheduler;
 
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
 import java.util.Set;
@@ -69,6 +70,7 @@ class HomePresenter implements BasePresenter {
      * to display it.
      */
     private void generateScheduleForCalendar() {
+        // todo: wrap this
         try {
             if (scheduler.generateSchedules()) {
                 Log.d("HomePresenter", "Conflict-free schedule generated!");
@@ -203,7 +205,6 @@ class HomePresenter implements BasePresenter {
     }
 
     public CharSequence[] getListOfAlternativeTimes(List<List<CourseComponent>> componentList) {
-        // todo: convert List<CourseComponent> to String
         CharSequence[] alternativeTimesArray = new CharSequence[componentList.size()];
         StringBuilder sb = new StringBuilder();
         sb.append("\n");
@@ -232,8 +233,33 @@ class HomePresenter implements BasePresenter {
     }
 
     public void saveCourses(List<List<CourseComponent>> list) {
-//        dataRepository.clearAllCourses();
-//        dataRepository.addAllCourses(list);
+
+    }
+
+    public List<List<CourseComponent>> getAlternativeSchedule(CourseComponent course, List<CourseComponent> selection, List<List<CourseComponent>> currSchedule) {
+        // todo: call young moula's shit
+        List<List<CourseComponent>> result = new ArrayList<>();
+        for (List<CourseComponent> courses : currSchedule) {
+            List<CourseComponent> subResult = new ArrayList<>();
+            for (CourseComponent c : courses) {
+                // curate new set of schedule by filtering out prev course section
+                if (!isCourseSameGroup(c, course)) {
+                    subResult.add(c);
+                }
+            }
+            result.add(subResult);
+        }
+        // update schedule with current section selection
+        result.add(selection);
+        return result;
+    }
+
+    private boolean isCourseSameGroup(CourseComponent prevCourse, CourseComponent currCourse) {
+        if (!prevCourse.getSubject().equals(currCourse.getSubject())) return false;
+        if (!prevCourse.getClassNumber().equals(currCourse.getClassNumber())) return false;
+        if (!prevCourse.getType().equals(currCourse.getType())) return false;
+        if (!prevCourse.getCatalogNumber().equals(currCourse.getCatalogNumber())) return false;
+        return true;
     }
 
     private String getTypeSpelling(String type) {
